backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0

# Data sources
yfinance==0.2.28
pytrends==4.9.2
fredapi==0.5.1

# Machine Learning
scikit-learn==1.3.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2

